% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_380}
\alias{ds_380}
\title{380-replicates set}
\format{
\subsection{\code{ds_380}}{

A data frame with 17,280 rows and 9 columns:
\describe{
\item{\code{well}}{Well identifier.}
\item{\code{replicate}}{Replicate identifier.}
\item{\code{dye}}{In all reactions the SYBR Green I master mix (Roche) was used,
so the value is always \code{"SYBR"}.}
\item{\code{target}}{Target identifier: MYCN.}
\item{\code{sample_type}}{Sample type: \code{"ntc"} (no template control) or \code{"std"} (standard).}
\item{\code{copies}}{Standard copy number: either \code{0} (for NTCs), or \code{150000} for standards.}
\item{\code{dilution}}{Dilution factor (the reciprocal of \code{copies}). Higher number
means greater dilution.}
\item{\code{cycle}}{PCR cycle.}
\item{\code{fluor}}{Raw fluorescence values.}
}
}
}
\source{
\url{https://medischebiologie.nl/wp-content/uploads/2019/02/qpcrdatamethods.zip}
}
\usage{
ds_380
}
\description{
The 380-replicates set is a dilution consisting of 150,000 molecules (see the
Admonition section below) of the MYCN oligonucleotide was created in 10 ng/µl
yeast tRNA carrier. qPCR amplifications were performed in 380 replicated 8 µl
reactions and quadruplicated reactions of the NTC sample were performed on
the same 384-well plate.

qPCR was done on a CFX 384 instrument (Bio-Rad). A 384-well qPCR plate was prepared using a
96-well head pipetting robot (Tecan Freedom Evo 150). qPCR amplifications
were performed in 8 µl containing:
\itemize{
\item 4 µl iQ SYBR Green Supermix (Bio-Rad);
\item 0.4 µl forward and 0.4 µl reverse primer (5 µM each);
\item 0.2 µl nuclease-free water;
\item 3 µl of standard oligonucleotide.
}

All reactions were performed in 384-well plates (Hard-Shell 384-well
microplates and Microseal B clear using adhesive seals (Bio-Rad).

The cycling conditions were comprised of 3 min polymerase activation at 95 °C
and 45 cycles of 15 s at 95 °C and 30 s at 60 °C followed by a dissociation
curve analysis from 60 to 95 °C.
\subsection{Admonition}{

The concentration information for the curves in this data set was seemingly
incorrectly indicated in the original publication (confirmed by email with
the senior author Prof. Jo Vandesompele). Instead of the 15,000 copies
reported in the original paper, the number of copies was most likely 150,000
copies (10x the value indicated).
}
}
\keyword{datasets}
