% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruijter-package.R
\docType{package}
\name{ruijter-package}
\alias{ruijter}
\alias{ruijter-package}
\title{ruijter: Technical Data Sets by Ruijter et al. (2013)}
\description{
The real-time quantitative polymerase chain reaction (qPCR) technical data sets by Ruijter et al. (2013) \doi{10.1016/j.ymeth.2012.08.011}: (i) the four-point 10-fold dilution series; (ii) 380 replicates; and (iii) the competimer data set. These three data sets can be used to benchmark qPCR methods. Original data set is available at \url{https://medischebiologie.nl/wp-content/uploads/2019/02/qpcrdatamethods.zip}. This package fixes incorrect annotations in the original data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ramiromagno/ruijter}
  \item \url{https://rmagno.eu/ruijter/}
  \item Report bugs at \url{https://github.com/ramiromagno/ruijter/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
